Renamed MacOS to Mac
Looks cleaner.
